pyproject.toml
[tool.poetry]
name = "doc-string"
version = "0.1.0"
description = ""
authors = ["Gerrit Schrader <gerrit.schrader@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.7.3"
docstring-parser = "^0.7.3"
pdoc3 = "^0.9.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
